Commit DAO Categoria, MetodoSpedizione,ItemOrdine

+++ src/main/java/it/unisa/nc26/digitronics/model/bean/MetodoSpedizione.java
@@ -25,7 +25,7 @@ public class MetodoSpedizione {
         return costo;
     }
 
-    public void setCosto(int costo) {
+    public void setCosto(double costo) {
         this.costo = costo;
     }
 }
